Extract `AnyString` nodes from `string/mod`
